# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  models.ErrorResponseUserSwagger:
    properties:
      code:
        example: 200
        type: integer
      error:
        example: Error details
        type: string
      message:
        example: 'Error message '
        type: string
      status:
        example: error or failed to validation request data or server error
        type: string
    type: object
...
  models.SuccessResponseUserSwagger:
    properties:
      code:
        example: 200
        type: integer
      data:
        properties:
          createdAt:
            example: "2021-07-01T00:00:00Z"
            type: string
          email:
            example: example@gmail.com
            type: string
          fullName:
            example: John Doe
            type: string
          isAdmin:
            example: false
            type: boolean
          isEmailVerified:
            example: false
            type: boolean
          isMobileVerified:
            example: false
            type: boolean
          profileImage:
            example: https://example.com/image.jpg
            type: string
          updatedAt:
            example: "2021-07-01T00:00:00Z"
            type: string
          userID:
            example: f47ac10b-58cc-4372-a567-0e02b2c3d479
            type: string
        type: object
      message:
        example: 'Success message '
        type: string
      status:
        example: success
        type: string
    type: object
  models.UserSwagger:
    description: Represents the details of a User
    properties:
      createdAt:
        example: "2021-07-01T00:00:00Z"
        type: string
      email:
        example: example@gmail.com
        type: string
      fullName:
        example: John Doe
        type: string
      isAdmin:
        example: false
        type: boolean
      isEmailVerified:
        example: false
        type: boolean
      isMobileVerified:
        example: false
        type: boolean
      password:
        example: Sdvdfgdf@234
        type: string
      profileImage:
        example: https://example.com/image.jpg
        type: string
      updatedAt:
        example: "2021-07-01T00:00:00Z"
        type: string
      userID:
        example: f47ac10b-58cc-4372-a567-0e02b2c3d479
        type: string
    type: object
...
  services.UserLoginData:
    properties:
      accessToken:
        type: string
      createdAt:
        type: string
      email:
        type: string
      fullName:
        type: string
      isAdmin:
        type: boolean
      isEmailVarified:
        type: boolean
      isMobileVarified:
        type: boolean
      profileImage:
        type: string
      refreshToken:
        type: string
      updatedAt:
        type: string
      userID:
        type: string
    type: object
  utils.AceesTokenAndRefreshToken:
    properties:
      access_token:
        type: string
      refresh_token:
        type: string
    type: object
host: https://carveo.site
info:
  contact:
    email: sanjaygupta7212@gmailcom
    name: Support Team
    url: https://carveo.com/support
  description: API documentation for Carevo, a car management system for tracking,
    servicing, and managing vehicles.
  termsOfService: https://carveo.com/terms
  title: Carevo Car Management API Documentation
  version: 1.0.0
paths:
...
  /userAuth/deleteUserProfile:
    delete:
      consumes:
      - application/json
      description: Delete User Profile
      produces:
      - application/json
      responses:
        "200":
          description: User profile deleted successfully
          schema:
            type: string
        "401":
          description: Unauthorized User
          schema:
            $ref: '#/definitions/models.ErrorResponseUserSwagger'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/models.ErrorResponseUserSwagger'
      security:
      - ApiKeyAuth: []
      summary: Delete User Profile
      tags:
      - User
  /userAuth/generateResetPasswordToken:
    post:
      consumes:
      - application/json
      description: Generate Reset Password Token
      parameters:
      - description: Email
        in: body
        name: email
        required: true
        schema:
          type: string
      produces:
      - application/json
      responses:
        "200":
          description: Reset password token generated successfully
          schema:
            type: string
        "400":
          description: Invalid input fields or JSON format
          schema:
            $ref: '#/definitions/models.ErrorResponseUserSwagger'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/models.ErrorResponseUserSwagger'
      summary: Generate Reset Password Token
      tags:
      - User
  /userAuth/getUserProfile:
    get:
      consumes:
      - application/json
      description: Get User Profile
      produces:
      - application/json
      responses:
        "200":
          description: User profile retrived successfully
          schema:
            $ref: '#/definitions/models.SuccessResponseUserSwagger'
        "401":
          description: Unauthorized User
          schema:
            $ref: '#/definitions/models.ErrorResponseUserSwagger'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/models.ErrorResponseUserSwagger'
      security:
      - ApiKeyAuth: []
      summary: Get User Profile
      tags:
      - User
  /userAuth/login:
    post:
      consumes:
      - application/json
      description: Login User
      parameters:
      - description: Email
        in: body
        name: email
        required: true
        schema:
          type: string
      - description: Password
        in: body
        name: password
        required: true
        schema:
          type: string
      produces:
      - application/json
      responses:
        "200":
          description: User logged in successfully
          schema:
            $ref: '#/definitions/services.UserLoginData'
        "400":
          description: Invalid input fields or JSON format
          schema:
            $ref: '#/definitions/models.ErrorResponseUserSwagger'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/models.ErrorResponseUserSwagger'
      summary: Login User
      tags:
      - User
  /userAuth/regenerateAccessAndRefreshToken:
    get:
      consumes:
      - application/json
      description: Regenerate Access and Refresh Token
      parameters:
      - description: Token
        in: query
        name: token
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Access and Refresh Token regenerated successfully
          schema:
            $ref: '#/definitions/utils.AceesTokenAndRefreshToken'
        "400":
          description: Invalid input fields or JSON format
          schema:
            $ref: '#/definitions/models.ErrorResponseUserSwagger'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/models.ErrorResponseUserSwagger'
      summary: Regenerate Access and Refresh Token
      tags:
      - User
  /userAuth/register:
    post:
      consumes:
      - application/json
      description: Register a new user
      parameters:
      - description: User Request
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/models.UserSwagger'
      produces:
      - application/json
      responses:
        "201":
          description: User created successfully
          schema:
            $ref: '#/definitions/models.SuccessResponseUserSwagger'
        "400":
          description: Invalid input fields or JSON format
          schema:
            $ref: '#/definitions/models.ErrorResponseUserSwagger'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/models.ErrorResponseUserSwagger'
      summary: Register a new user
      tags:
      - User
  /userAuth/updatePassword:
    patch:
      consumes:
      - application/json
      description: Update Password
      parameters:
      - description: Password
        in: body
        name: password
        required: true
        schema:
          type: string
      - description: Email
        in: body
        name: email
        required: true
        schema:
          type: string
      produces:
      - application/json
      responses:
        "200":
          description: Password updated successfully
          schema:
            type: string
        "400":
          description: Invalid input fields or JSON format
          schema:
            $ref: '#/definitions/models.ErrorResponseUserSwagger'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/models.ErrorResponseUserSwagger'
      summary: Update Password
      tags:
      - User
  /userAuth/updateUserProfile:
    patch:
      consumes:
      - application/json
      description: Update User Profile
      parameters:
      - description: User Request
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/models.UserSwagger'
      produces:
      - application/json
      responses:
        "200":
          description: User profile updated successfully
          schema:
            $ref: '#/definitions/models.SuccessResponseUserSwagger'
        "400":
          description: Invalid input fields or JSON format
          schema:
            $ref: '#/definitions/models.ErrorResponseUserSwagger'
        "401":
          description: Unauthorized User
          schema:
            $ref: '#/definitions/models.ErrorResponseUserSwagger'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/models.ErrorResponseUserSwagger'
      security:
      - ApiKeyAuth: []
      summary: Update User Profile
      tags:
      - User
  /userAuth/validateResetPasswordToken:
    get:
      consumes:
      - application/json
      description: Validate Reset Password Token
      parameters:
      - description: Token
        in: query
        name: token
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Reset password token verified successfully
          schema:
            type: string
        "400":
          description: Invalid input fields or JSON format
          schema:
            $ref: '#/definitions/models.ErrorResponseUserSwagger'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/models.ErrorResponseUserSwagger'
      summary: Validate Reset Password Token
      tags:
      - User
  /userAuth/verifyEmail:
    get:
      consumes:
      - application/json
      description: Verify Email
      parameters:
      - description: Token
        in: query
        name: token
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Email verified successfully
          schema:
            type: string
        "400":
          description: Invalid input fields or JSON format
          schema:
            $ref: '#/definitions/models.ErrorResponseUserSwagger'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/models.ErrorResponseUserSwagger'
      summary: Verify Email
      tags:
      - User
securityDefinitions:
  CarveoAPIKey:
    description: 'JWT token required to access endpoints. Add the token in the "Authorization"
      header in the format: Bearer <JWT_TOKEN>.'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"
...